requirements.txt
Django == 5.0.2
django-extensions==3.2.3
djangorestframework==3.14.0
django-bootstrap4

